modified queue and add reservation table

+++ inventory/internal/ports/inventory_repository.go
@@ -5,5 +5,5 @@ import (
 )
 
 type InventoryRepoPort interface {
-	GetProductQuantity(ctx context.Context, productCode string) (int64, error)
+	GetProductQuantity(ctx context.Context, productCode string) (int32, error)
 }
